Website v1
What there is: searchable by county, reasonable plots.  Boring, but functional colors.  Fully deployable static website which chunks all of our data by county.  URLs are shareable (if I see a plot I want to share on e.g. social media, I can just copy+paste the URL and they'll get the exact same view).

What needs to be fixed: the CountyOverTime javascript file is unreasonably large (2MB).  I presume this is because e.g. Chart.js is bundled with it, but I intend to look into that.

What there is not: testing.  A one-click export of all data to XLS like we talked about.  A cron pipeline which updates the site daily.